backend/BACKEND/requirements.txt
altair==6.0.0
asgiref==3.11.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
lime==0.2.0.1
MarkupSafe==3.0.3
matplotlib==3.10.8
mysqlclient==2.2.8
narwhals==2.16.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pillow==12.1.1
protobuf==6.33.5
pyarrow==23.0.0
pydeck==0.9.1
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
smmap==5.0.2
sqlparse==0.5.5
streamlit==1.54.0
tenacity==9.1.4
threadpoolctl==3.6.0
tifffile==2026.1.28
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
